pyproject.toml
[project]
name = "sales_agent"
version = "0.1.0"
description = "SOTA Sales Agent Testing Framework"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "python-dotenv",
    "openai",
    "azure-identity",
    "pandas",
]

[project.scripts]
sales-agent = "sales_agent.main:main"

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.ruff.lint]
select = ["E", "F", "I"]
ignore = []

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["src"]
